improve composition logic
